+++ src/main/java/com/LetMeDoWith/LetMeDoWith/presentation/task/controller/TaskController.java
@@ -1,7 +1,6 @@
 package com.LetMeDoWith.LetMeDoWith.presentation.task.controller;
 
 import com.LetMeDoWith.LetMeDoWith.application.task.dto.RetrieveTasksResult;
-import com.LetMeDoWith.LetMeDoWith.application.task.dto.RetrieveTodoTaskResult;
 import com.LetMeDoWith.LetMeDoWith.application.task.service.RetrieveTaskService;
 import com.LetMeDoWith.LetMeDoWith.common.annotation.ApiErrorResponse;
 import com.LetMeDoWith.LetMeDoWith.common.annotation.ApiErrorResponses;
@@ -10,7 +9,6 @@ import com.LetMeDoWith.LetMeDoWith.common.dto.ResponseDto;
 import com.LetMeDoWith.LetMeDoWith.common.exception.status.FailResponseStatus;
 import com.LetMeDoWith.LetMeDoWith.common.util.ResponseUtil;
 import com.LetMeDoWith.LetMeDoWith.presentation.task.dto.RetrieveTasksResDto;
-import com.LetMeDoWith.LetMeDoWith.presentation.task.dto.RetrieveTodoTaskResDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.time.Month;
@@ -18,7 +16,6 @@ import java.time.Year;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -43,14 +40,5 @@ public class TaskController {
         RetrieveTasksResult result = retrieveTaskService.retrieveMonthTasks(Year.of(year), Month.of(month));
 
         return ResponseUtil.createSuccessResponse(RetrieveTasksResDto.from(result));
-    }
-
-    @Operation(summary = "투두 태스크 단일 조회", description = "1개의 TodoTask 정보를 조회합니다.")
-    @ApiSuccessResponse(description = "1개의 TodoTask 정보 및 루틴 정보를 반환")
-    @GetMapping("/todo/{todoTaskId}")
-    public ResponseEntity<ResponseDto<RetrieveTodoTaskResDto>> retrieveSingleTodoTask(@PathVariable Long todoTaskId) {
-        RetrieveTodoTaskResult result = retrieveTaskService.retrieveTodoTask(todoTaskId);
-
-        return ResponseUtil.createSuccessResponse(RetrieveTodoTaskResDto.from(result));
     }
 }
